Create codex test file
